refactor(ponytail): fix getMany bug, cleanup unused storage abstractions & drop deprecated DB columns
- Fix getMany() returning Map instead of array, fixing silent undefined destructuring in contexts/settings
- Fix DataSyncer calculationMethod event trigger to avoid overriding local state on login
- Fix FCM auto-init to include prayerPreferences and full location
- Delete unused StorageFactory and error classes
- Delete unnecessary barrel index files (analytics, habits, quran, islamic-content)
- Add and apply DB migration 0016 to drop unused columns (push_subscription.user_location, ramadhan_taraweh_log.is_qiyamul_lail, ramadhan_daily_log.fardhu_location)
