--- src/lib/analytics/index.ts
@@ -1,3 +1,0 @@
-export * from './analytics';
-export * from './analytics-utils';
-export * from './server-analytics';
