--- src/lib/quran/index.ts
@@ -1,7 +1,0 @@
-export * from './kemenag-api';
-export * from './optimize-quran';
-export * from './quran-segments-api';
-export * from './quran-utils';
-export * from './surahData';
-export * from './tafsir-api';
-export * from './bookmark-storage';
